% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLIHC450k.R
\name{loadLIHC450k}
\alias{loadLIHC450k}
\title{use BiocFileCache discipline to acquire lihc450K SummarizedExperiment}
\usage{

  loadLIHC450k(remoteSEPath = "https://s3.amazonaws.com/biocfound-tcga/lihc450k.rds",
  remoteHSDSPath = "http://52.4.181.237:5101",
  hostPath = "/home/stvjc/lihc450k.h5", cache = BiocFileCache())
}
\arguments{
\item{remoteSEPath}{character(1) identifying remote RDS with assay-free SummarizedExperiment instance}

\item{remoteHSDSPath}{character(1) identifying HSDS server/port}

\item{hostPath}{character(1) the data 'host' -- looks strange with '/home' default but it is OK}

\item{cache}{defaults to BiocFileCache() value}
}
\description{
use BiocFileCache discipline to acquire lihc450K SummarizedExperiment
}
\note{
The RDS for the SummarizedExperiment is in an AWS S3 bucket.
This function will check local cache for the data and will download
to cache if not found.  That download is a one-time operation for
any given value of \code{cache}.  Note further that the `remoteHSDSPath`
is provided courtesy of the HDF Group in early 2018 and may not be persistent.
}
\examples{
lim = loadLIHC450k()
assay(lim)
assay(lim)[1,429]
# 'confirm' with restfulSE interface to ISB-CGC
# cgcConnection() \%>\% tbl("DNA_Methylation_chr16") \%>\% 
#    filter( ParticipantBarcode=="TCGA-ZS-A9CG", Probe_Id == "cg00000029") \%>\% 
#    glimpse()
# Observations: ??
# Running job \\:  1s:18.9 gigabytes processed
# Variables: 8                                                                  
# $ ParticipantBarcode   <chr> "TCGA-ZS-A9CG"
# $ SampleBarcode        <chr> "TCGA-ZS-A9CG-01A"
# $ SampleTypeLetterCode <chr> "TP"
# $ AliquotBarcode       <chr> "TCGA-ZS-A9CG-01A-11D-A36Y-05"
# $ Platform             <chr> "HumanMethylation450"
# $ Study                <chr> "LIHC"
# $ Probe_Id             <chr> "cg00000029"
# $ Beta_Value           <dbl> 0.65
#
}
